Add authorization

+++ app/controllers/api/companies_controller.rb
@@ -11,6 +11,8 @@ module Api
     end
 
     def create
+      return error_message if check_user || find_user.role != 'admin'
+
       company = Company.new(permitted_params)
 
       if company.save
@@ -21,6 +23,8 @@ module Api
     end
 
     def update
+      return error_message if check_user || find_user.role != 'admin'
+
       company = Company.find(params[:id])
       if company.update(permitted_params)
         render json: CompanySerializer.render(company, root: 'company'), status: :ok
@@ -30,6 +34,8 @@ module Api
     end
 
     def destroy
+      return error_message if check_user || find_user.role != 'admin'
+
       company = Company.find(params[:id])
       company.destroy
       head :no_content
@@ -39,6 +45,20 @@ module Api
 
     def permitted_params
       params.require(:company).permit(:name)
+    end
+
+    def error_message
+      render json: { errors: { token: ['is invalid'] } }, status: :unauthorized
+    end
+
+    def find_user
+      token = request.headers['Authorization']
+      User.find_by(token: token)
+    end
+
+    def check_user
+      token = request.headers['Authorization']
+      User.find_by(token: token).nil?
     end
   end
 end
