+++ app/controllers/api/users_controller.rb
@@ -1,7 +1,7 @@
 module Api
   class UsersController < ApplicationController
     def index
-      return error_message if check_user
+      return error_message if check_user || find_user.role != 'admin'
 
       render json: UserSerializer.render(User.all, root: 'users'), status: :ok
     end
@@ -10,6 +10,8 @@ module Api
       return error_message if check_user
 
       user = User.find(params[:id])
+      return error_message if find_user.id != user.id && find_user.role != 'admin'
+
       render json: UserSerializer.render(user, root: 'user'), status: :ok
     end
 
@@ -27,17 +29,17 @@ module Api
       return error_message if check_user
 
       user = User.find(params[:id])
-      if user.update(permitted_params)
-        render json: UserSerializer.render(user, root: 'user'), status: :ok
-      else
-        render json: { errors: user.errors }, status: :bad_request
-      end
+      return error_message if find_user.id != user.id && find_user.role != 'admin'
+
+      update_user(user)
     end
 
     def destroy
       return error_message if check_user
 
       user = User.find(params[:id])
+      return error_message if find_user.id != params[:id] && find_user.role != 'admin'
+
       user.destroy
       head :no_content
     end
@@ -55,6 +57,19 @@ module Api
 
     def error_message
       render json: { errors: { token: ['is invalid'] } }, status: :unauthorized
+    end
+
+    def find_user
+      token = request.headers['Authorization']
+      User.find_by(token: token)
+    end
+
+    def update_user(user)
+      if user.update(permitted_params)
+        render json: UserSerializer.render(user, root: 'user'), status: :ok
+      else
+        render json: { errors: user.errors }, status: :bad_request
+      end
     end
   end
 end
